requirements.txt
pyrama
numpy
biopython
pandas
scipy
